% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_between.R
\name{area_between}
\alias{area_between}
\title{A function to calculate area between two profiles}
\usage{
area_between(s1, s2 = NA, h = NA, sigma_z = c(NA, NA), sigma_l = c(NA,
  NA))
}
\arguments{
\item{s1}{tibble with columns l and z describing first profile}

\item{s2}{tibble with columns l and z describing second profile}

\item{h}{if provided by user, the second profile is supposed to be horizontal, with constant height=h (defaults to NA)}

\item{sigma_z}{a vector of length 1 or 2 providing an estimate of the error in measures of height z1 and z2}

\item{sigma_l}{a vector of length 1 or 2 providing an estimate of the error in measures of longitudinal coordinates l1 and l2z1 and z2}
}
\value{
area

area_by_type

sigma_area

data
}
\description{
This function takes two profiles defined as (l1,z1) and (l2,z2) as inputs and calculates area between them
}
\examples{
s1 <- tibble(l=c(1,3,5,6,9),
             z=c(1,2,3,2.5,5))
s2 <- tibble(l=c(0.5,2.5,4,6,8),
             z=c(3,1,2,4,3))
area_between(s1,s2)
area_between(s1,h=2)
}
