% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_transect.R
\name{clean_transect}
\alias{clean_transect}
\title{A function that cleans and formats sr.}
\usage{
clean_transect(sr)
}
\arguments{
\item{sr}{tibble with columns l and z describing first profile}
}
\value{
a clean sr data with variables id, l, and z (arranged by ascending l)
}
\description{
A function that cleans and formats sr.
}
