% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_transect.R
\name{cut_series}
\alias{cut_series}
\title{A function that returns the part of sr delimited by a and b.}
\usage{
cut_series(sr, channels)
}
\arguments{
\item{sr}{tibble with columns l and z describing profile}

\item{channels}{the channel-describing data to use to cut the data into sub-series}
}
\value{
the part of sr that is comprised between a and b
}
\description{
A function that returns the part of sr delimited by a and b.
}
\examples{
data(s1)
channels=get_channels(s1,hmin=1,hmax=4.5)
cut_series(s1, channels)
}
