% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sections.R
\name{find_section}
\alias{find_section}
\title{Describe levee-to-levee section in which a local minimum with z=z_min falls.}
\usage{
find_section(rtib, tibm, sr, hmin)
}
\arguments{
\item{rtib}{tibble describing local minimum (id and z)}

\item{tibm}{tibble with local maxima}

\item{sr}{tibble with columns l and z describing profile}

\item{hmin}{the minimum value for a local maximum to be delimiting a section (z > z_min + hmin)}
}
\value{
a tibble describing the section.
}
\description{
Describe levee-to-levee section in which a local minimum with z=z_min falls.
}
