% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channels.R
\name{get_channel_from_section}
\alias{get_channel_from_section}
\title{A function that returns channels delimited by one local maxima (on one side).}
\usage{
get_channel_from_section(section, sr)
}
\arguments{
\item{section}{tibble with a section's descriptors (one row-tibble)}

\item{sr}{sr tibble with columns l and z describing profile}
}
\value{
descriptors of the channel corresponding to section
}
\description{
A function that returns channels delimited by one local maxima (on one side).
}
