% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channels.R
\name{get_channels}
\alias{get_channels}
\title{A function that returns sections and channels}
\usage{
get_channels(sr, hmin, hmax, type = "bankfull")
}
\arguments{
\item{sr}{tibble with columns l and z describing profile}

\item{hmin}{the minimum height (relative to local minimum of z) of levees-defining local maxima}

\item{hmax}{the maximum height (relative to minimum of z) of a channel-defining local minima}

\item{type}{whether to return section or bankfull channel ("section" or "bankfull"). Defaults to "bankfull"}
}
\value{
channels a tibble with sections and channels descriptors
}
\description{
A function that returns sections and channels
}
\examples{
data(s1)
channels=get_channels(s1, hmin=1, hmax=5.5)
}
