% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sections.R
\name{get_sections}
\alias{get_sections}
\title{A function that returns sections delimited by local maxima with height > hmin.}
\usage{
get_sections(sr, hmin, hmax)
}
\arguments{
\item{sr}{tibble with columns l and z describing first profile}

\item{hmin}{the minimum height (relative to local minimum of z) of levees-defining local maxima}

\item{hmax}{the maximum height (relative to minimum of z) of a channel-defining local minima}
}
\value{
hmin
}
\description{
A function that returns sections delimited by local maxima with height > hmin.
}
\examples{
data(s1)
get_sections(s1, hmin=1,hmax=5)
}
