% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channels.R
\name{plot_channels}
\alias{plot_channels}
\title{A function that plots channels and sections.}
\usage{
plot_channels(channels, sr)
}
\arguments{
\item{channels}{a table with channels and sections coordinates, as returned by get_channels}

\item{sr}{tibble with columns l and z describing profile}
}
\value{
a plot
}
\description{
A function that plots channels and sections.
}
\examples{
data(s1)
channels=get_channels(s1, hmin=1, hmax=5.5)
plot_channels(channels,s1)
}
