% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_on_AB.R
\name{project_on_AB}
\alias{project_on_AB}
\title{A function to project data on a line defined by two points A & B}
\usage{
project_on_AB(dat, datAB, datOg = datAB[1, ])
}
\arguments{
\item{dat}{tibble with coordinates columns x and y}

\item{datAB}{tibble with two rows (points A and B) and coordinates columns x and y}

\item{datOg}{point that defines the origin (point with l=0) of the series}
}
\value{
xp x coordinate of points projected on the transect

yp y coordinate of points projected on the transect

xt x coordinate along the transect (new coordinate x)

yt y coordinate corresponding to the distance to the transect (new coordinate y)
}
\description{
A function to project data on a line defined by two points A & B
}
\examples{
dat=tibble(x=c(866142.3, 866149.9, 866143.7),
           y=c(84496.97, 84496.93, 84498.50))
datAB=tibble(x=c(866141.3,866148.7),
             y=c(84498.97, 84496.50))
dat=project_on_AB(dat,datAB)
ggplot(dat,  aes(x=x,y=y))+
  geom_point(col="red")+
  geom_line(data=datAB, aes(x=x,y=y))+
  geom_point(data=datAB)+
  geom_point(data=dat,col="blue", aes(x=xp,y=yp))+
  coord_fixed(ratio=1)
}
