% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regroup_channels.R
\name{regroup_channels}
\alias{regroup_channels}
\title{A function that cleans and formats sr.}
\usage{
regroup_channels(channels, by_id)
}
\arguments{
\item{channels}{the table describing channels}

\item{by_id}{the channel id (defines how the regrouping is done)}

\item{sr}{tibble with columns l and z describing first profile}
}
\value{
a clean sr data with variables id, l, and z (arranged by ascending l)
}
\description{
A function that cleans and formats sr.
}
