% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_series.R
\name{cross_series}
\alias{cross_series}
\title{A function to calculate area between two profiles}
\usage{
cross_series(s1, s2, h = NA)
}
\arguments{
\item{s1}{tibble with columns l and z describing first profile}

\item{s2}{tibble with columns l and z describing second profile}

\item{h}{if provided by user, the second profile is supposed to be horizontal, with constant height=h (defaults to NA)}
}
\value{
area

area_by_type

data
}
\description{
This function takes two profiles defined as (l1,z1) and (l2,z2) as inputs and calculates area between them
}
\examples{
l1=c(1,3,5,6,9)
z1=c(1,2,3,2.5,5)
l2=c(0.5,2.5,4,6,8)
z2=c(3,1,2,4,3)
cross_series(l1,z1,l2,z2)
cross_series(l1,z1,h=2)
}
