% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_series.R
\name{cross_series}
\alias{cross_series}
\title{This function crosses two (l,z) series and provides a tibble regarding all points of interest (observations, intersections, interpolations)}
\usage{
cross_series(s1, s2, h = NA)
}
\arguments{
\item{s1}{tibble with columns l and z describing first profile}

\item{s2}{tibble with columns l and z describing second profile}

\item{h}{if provided by user, the second profile is supposed to be horizontal, with constant height=h (defaults to NA)}
}
\value{
a tibble providing info on all points of interest in the crossing of the two series.
}
\description{
This function crosses two (l,z) series and provides a tibble regarding all points of interest (observations, intersections, interpolations)
}
\examples{
data(s1)
data(s2)
cross_series(s1,s2)
cross_series(s1,h=2)
}
