% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_area.R
\name{plot_area}
\alias{plot_area}
\title{A function to plot the area between two profiles}
\usage{
plot_area(result_area, distinguish_type = TRUE,
  show_uncertainty = FALSE)
}
\arguments{
\item{distinguish_type}{whether or not to distinguish between upper area and lower area}

\item{show_uncertainty}{whether or not to show the sources of error}

\item{area_data}{the result of function area_between() or area_uncertainty()}
}
\value{
a plot showing the area difference between series
}
\description{
This function takes as input the result of function area_between()
}
\examples{
l1=c(1,3,5,6,9)
z1=c(1,2,3,2.5,5)
l2=c(0.5,2.5,4,6,8)
z2=c(3,1,2,4,3)
result_area <- area_between(l1,z1,l2,z1)
plot_area(result_area)
}
